chore: move shared touch helper
