--- automation/src/main/java/info/nightscout/androidaps/plugins/general/automation/dragHelpers/SimpleItemTouchHelperCallback.kt
@@ -1,37 +1,0 @@
-package info.nightscout.androidaps.plugins.general.automation.dragHelpers
-
-import androidx.recyclerview.widget.ItemTouchHelper
-import androidx.recyclerview.widget.RecyclerView
-import info.nightscout.androidaps.plugins.general.automation.AutomationFragment
-
-class SimpleItemTouchHelperCallback : ItemTouchHelper.SimpleCallback(ItemTouchHelper.UP or ItemTouchHelper.DOWN or ItemTouchHelper.START or ItemTouchHelper.END, 0) {
-
-    override fun isLongPressDragEnabled(): Boolean {
-        return false
-    }
-
-    override fun onMove(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder, target: RecyclerView.ViewHolder): Boolean {
-        val adapter = recyclerView.adapter as AutomationFragment.EventListAdapter
-        val from = viewHolder.layoutPosition
-        val to = target.layoutPosition
-        adapter.onItemMove(from, to)
-        adapter.notifyItemMoved(from, to)
-
-        return true
-    }
-
-    override fun onSwiped(viewHolder: RecyclerView.ViewHolder, direction: Int) {}
-
-    override fun onSelectedChanged(viewHolder: RecyclerView.ViewHolder?, actionState: Int) {
-        super.onSelectedChanged(viewHolder, actionState)
-        if (actionState == ItemTouchHelper.ACTION_STATE_DRAG) {
-            viewHolder?.itemView?.alpha = 0.5f
-        }
-    }
-
-    override fun clearView(recyclerView: RecyclerView, viewHolder: RecyclerView.ViewHolder) {
-        super.clearView(recyclerView, viewHolder)
-        viewHolder.itemView.alpha = 1.0f
-        (recyclerView.adapter as AutomationFragment.EventListAdapter).onDrop()
-    }
-}
